+++ runtime/server_http_request.go
@@ -58,7 +58,7 @@ type ServerHTTPRequest struct {
 	Header       Header
 
 	// logger logs entries with default fields that contains request meta info
-	logger Logger
+	contextLogger ContextLogger
 	// scope emit metrics with default tags that contains request meta info
 	scope       tally.Scope
 	jsonWrapper jsonwrapper.JSONWrapper
@@ -105,21 +105,21 @@ func NewServerHTTPRequest(
 	httpRequest := r.WithContext(ctx)
 
 	scope := endpoint.scope.Tagged(scopeTags)
-	logger := endpoint.logger.With(logFields...)
+	logger := endpoint.contextLogger
 
 	req := &ServerHTTPRequest{
-		httpRequest:  httpRequest,
-		queryValues:  nil,
-		tracer:       endpoint.tracer,
-		EndpointName: endpoint.EndpointName,
-		HandlerName:  endpoint.HandlerName,
-		URL:          httpRequest.URL,
-		Method:       httpRequest.Method,
-		Params:       params,
-		Header:       NewServerHTTPHeader(r.Header),
-		logger:       logger,
-		scope:        scope,
-		jsonWrapper:  endpoint.JSONWrapper,
+		httpRequest:   httpRequest,
+		queryValues:   nil,
+		tracer:        endpoint.tracer,
+		EndpointName:  endpoint.EndpointName,
+		HandlerName:   endpoint.HandlerName,
+		URL:           httpRequest.URL,
+		Method:        httpRequest.Method,
+		Params:        params,
+		Header:        NewServerHTTPHeader(r.Header),
+		contextLogger: logger,
+		scope:         scope,
+		jsonWrapper:   endpoint.JSONWrapper,
 	}
 
 	req.res = NewServerHTTPResponse(w, req)
@@ -141,7 +141,7 @@ func (req *ServerHTTPRequest) StartTime() time.Time {
 func (req *ServerHTTPRequest) start() {
 	if req.started {
 		/* coverage ignore next line */
-		req.logger.Error(
+		req.contextLogger.Error(req.Context(),
 			"Cannot start ServerHTTPRequest twice",
 			zap.String("path", req.URL.Path),
 		)
@@ -164,7 +164,7 @@ func (req *ServerHTTPRequest) start() {
 		if err != nil {
 			if err != opentracing.ErrSpanContextNotFound {
 				/* coverage ignore next line */
-				req.logger.Warn("Error Extracting Trace Headers", zap.Error(err))
+				req.contextLogger.Warn(req.Context(), "Error Extracting Trace Headers", zap.Error(err))
 			}
 			span = req.tracer.StartSpan(opName, urlTag, MethodTag)
 		} else {
@@ -179,7 +179,7 @@ func (req *ServerHTTPRequest) CheckHeaders(headers []string) bool {
 	for _, headerName := range headers {
 		_, ok := req.Header.Get(headerName)
 		if !ok {
-			req.logger.Warn("Got request without mandatory header",
+			req.contextLogger.Warn(req.Context(), "Got request without mandatory header",
 				zap.String("headerName", headerName),
 			)
 
@@ -225,7 +225,7 @@ func (req *ServerHTTPRequest) parseQueryValues() bool {
 
 	values, err := url.ParseQuery(req.httpRequest.URL.RawQuery)
 	if err != nil {
-		req.logger.Warn("Got request with invalid query string", zap.Error(err))
+		req.contextLogger.Warn(req.Context(), "Got request with invalid query string", zap.Error(err))
 
 		if !req.parseFailed {
 			req.res.SendErrorString(
@@ -700,7 +700,7 @@ func (req *ServerHTTPRequest) CheckQueryValue(key string) bool {
 
 	values := req.queryValues[key]
 	if len(values) == 0 {
-		req.logger.Warn("Got request with missing query string value",
+		req.contextLogger.Warn(req.Context(), "Got request with missing query string value",
 			zap.String("expectedKey", key),
 		)
 		if !req.parseFailed {
@@ -753,7 +753,7 @@ func (req *ServerHTTPRequest) ReadAll() ([]byte, bool) {
 	}
 	rawBody, err := ioutil.ReadAll(req.httpRequest.Body)
 	if err != nil {
-		req.logger.Error("Could not read request body", zap.Error(err))
+		req.contextLogger.Error(req.Context(), "Could not read request body", zap.Error(err))
 		if !req.parseFailed {
 			req.res.SendError(500, "Could not read request body", err)
 			req.parseFailed = true
@@ -770,7 +770,7 @@ func (req *ServerHTTPRequest) UnmarshalBody(
 ) bool {
 	err := req.jsonWrapper.Unmarshal(rawBody, body)
 	if err != nil {
-		req.logger.Warn("Could not parse json", zap.Error(err))
+		req.contextLogger.Warn(req.Context(), "Could not parse json", zap.Error(err))
 		if !req.parseFailed {
 			req.res.SendError(400, "Could not parse json: "+err.Error(), err)
 			req.parseFailed = true
@@ -799,7 +799,7 @@ func (req *ServerHTTPRequest) GetSpan() opentracing.Span {
 
 // LogAndSendQueryError handles parse failure of query params by logging the issue and returning a 400 to the requestor
 func (req *ServerHTTPRequest) LogAndSendQueryError(err error, expected, key, value string) {
-	req.logger.Warn("Got request with invalid query string types",
+	req.contextLogger.Warn(req.Context(), "Got request with invalid query string types",
 		zap.String("expected", expected),
 		zap.String("actual", value),
 		zap.String("key", key),