+++ runtime/server_http_response.go
@@ -47,10 +47,10 @@ type ServerHTTPResponse struct {
 	pendingBodyBytes  []byte
 	pendingBodyObj    interface{}
 	pendingStatusCode int
-	logger            Logger
+	contextLogger     ContextLogger
 	scope             tally.Scope
 	jsonWrapper       jsonwrapper.JSONWrapper
-	err               error
+	Err               error
 }
 
 // NewServerHTTPResponse is helper function to alloc ServerHTTPResponse
@@ -62,7 +62,7 @@ func NewServerHTTPResponse(
 		Request:        req,
 		StatusCode:     200,
 		responseWriter: w,
-		logger:         req.logger,
+		contextLogger:  req.contextLogger,
 		scope:          req.scope,
 		jsonWrapper:    req.jsonWrapper,
 	}
@@ -73,7 +73,7 @@ func (res *ServerHTTPResponse) finish(ctx context.Context) {
 	logFields := GetLogFieldsFromCtx(ctx)
 	if !res.Request.started {
 		/* coverage ignore next line */
-		res.logger.Error(
+		res.contextLogger.Error(ctx,
 			"Forgot to start server response",
 			append(logFields, zap.String("path", res.Request.URL.Path))...,
 		)
@@ -82,7 +82,7 @@ func (res *ServerHTTPResponse) finish(ctx context.Context) {
 	}
 	if res.finished {
 		/* coverage ignore next line */
-		res.logger.Error(
+		res.contextLogger.Error(ctx,
 			"Finished a server response multiple times",
 			append(logFields, zap.String("path", res.Request.URL.Path))...,
 		)
@@ -100,7 +100,7 @@ func (res *ServerHTTPResponse) finish(ctx context.Context) {
 	tagged.Timer(endpointLatency).Record(delta)
 	tagged.Histogram(endpointLatencyHist, tally.DefaultBuckets).RecordDuration(delta)
 	if !known {
-		res.logger.Error(
+		res.contextLogger.Error(ctx,
 			"Unknown status code",
 			append(logFields, zap.Int("UnknownStatusCode", res.StatusCode))...,
 		)
@@ -108,10 +108,10 @@ func (res *ServerHTTPResponse) finish(ctx context.Context) {
 		tagged.Counter(endpointStatus).Inc(1)
 	}
 
-	logFn := res.logger.Debug
+	logFn := res.contextLogger.Debug
 	if !known || res.StatusCode >= 400 && res.StatusCode < 600 {
 		tagged.Counter(endpointAppErrors).Inc(1)
-		logFn = res.logger.Warn
+		logFn = res.contextLogger.Warn
 	}
 
 	span := res.Request.GetSpan()
@@ -119,7 +119,7 @@ func (res *ServerHTTPResponse) finish(ctx context.Context) {
 		span.Finish()
 	}
 
-	logFn(
+	logFn(ctx,
 		fmt.Sprintf("Finished an incoming server HTTP request with %d status code", res.StatusCode),
 		append(logFields, serverHTTPLogFields(res.Request, res)...)...,
 	)
@@ -145,11 +145,11 @@ func serverHTTPLogFields(req *ServerHTTPRequest, res *ServerHTTPResponse) []zapc
 		}
 	}
 
-	if res.err != nil {
-		fields = append(fields, zap.Error(res.err))
+	if res.Err != nil {
+		fields = append(fields, zap.Error(res.Err))
 
-		cause := errors.Cause(res.err)
-		if cause != nil && cause != res.err {
+		cause := errors.Cause(res.Err)
+		if cause != nil && cause != res.Err {
 			fields = append(fields, zap.NamedError("errorCause", cause))
 		}
 	}
@@ -170,7 +170,7 @@ func (res *ServerHTTPResponse) SendErrorString(
 func (res *ServerHTTPResponse) SendError(
 	statusCode int, errMsg string, errCause error,
 ) {
-	res.err = errCause
+	res.Err = errCause
 	res.WriteJSONBytes(statusCode, nil,
 		[]byte(`{"error":"`+errMsg+`"}`),
 	)
@@ -207,15 +207,16 @@ func (res *ServerHTTPResponse) WriteJSONBytes(
 
 // MarshalResponseJSON serializes a json serializable into bytes
 func (res *ServerHTTPResponse) MarshalResponseJSON(body interface{}) []byte {
+	ctx := res.Request.Context()
 	if body == nil {
 		res.SendError(500, "Could not serialize json response", errors.New("No Body JSON"))
-		res.logger.Error("Could not serialize nil pointer body")
+		res.contextLogger.Error(ctx, "Could not serialize nil pointer body")
 		return nil
 	}
 	bytes, err := res.jsonWrapper.Marshal(body)
 	if err != nil {
 		res.SendError(500, "Could not serialize json response", err)
-		res.logger.Error("Could not serialize json response", zap.Error(err))
+		res.contextLogger.Error(ctx, "Could not serialize json response", zap.Error(err))
 		return nil
 	}
 	return bytes
@@ -280,7 +281,7 @@ func (res *ServerHTTPResponse) PeekBody(
 func (res *ServerHTTPResponse) flush(ctx context.Context) {
 	if res.flushed {
 		/* coverage ignore next line */
-		res.logger.Error(
+		res.contextLogger.Error(ctx,
 			"Flushed a server response multiple times",
 			zap.String("path", res.Request.URL.Path),
 		)
@@ -304,7 +305,7 @@ func (res *ServerHTTPResponse) writeBytes(bytes []byte) {
 	_, err := res.responseWriter.Write(bytes)
 	if err != nil {
 		/* coverage ignore next line */
-		res.logger.Error(
+		res.contextLogger.Error(res.Request.Context(),
 			"Could not write string to resp body",
 			zap.Error(err),
 		)