+++ runtime/tchannel_server.go
@@ -54,11 +54,11 @@ type TChannelEndpoint struct {
 // TChannelRouter handles incoming TChannel calls and routes them to the matching TChannelHandler.
 type TChannelRouter struct {
 	sync.RWMutex
-	registrar tchannel.Registrar
-	endpoints map[string]*TChannelEndpoint
-	logger    *zap.Logger
-	scope     tally.Scope
-	extractor ContextExtractor
+	registrar     tchannel.Registrar
+	endpoints     map[string]*TChannelEndpoint
+	contextLogger ContextLogger
+	scope         tally.Scope
+	extractor     ContextExtractor
 
 	requestUUIDHeaderKey string
 }
@@ -103,11 +103,11 @@ func NewTChannelEndpointWithPostResponseCB(
 // NewTChannelRouter returns a TChannel router that can serve thrift services over TChannel.
 func NewTChannelRouter(registrar tchannel.Registrar, g *Gateway) *TChannelRouter {
 	return &TChannelRouter{
-		registrar: registrar,
-		endpoints: map[string]*TChannelEndpoint{},
-		logger:    g.Logger,
-		scope:     g.RootScope,
-		extractor: g.ContextExtractor,
+		registrar:     registrar,
+		endpoints:     map[string]*TChannelEndpoint{},
+		contextLogger: g.ContextLogger,
+		scope:         g.RootScope,
+		extractor:     g.ContextExtractor,
 
 		requestUUIDHeaderKey: g.requestUUIDHeaderKey,
 	}
@@ -134,7 +134,7 @@ func (s *TChannelRouter) Register(e *TChannelEndpoint) error {
 func (s *TChannelRouter) Handle(ctx context.Context, call *tchannel.InboundCall) {
 	method := call.MethodString()
 	if sep := strings.Index(method, "::"); sep == -1 {
-		s.logger.Error("Handle got call for which does not match the expected call format", zap.String(logFieldRequestMethod, method))
+		s.contextLogger.Error(ctx, "Handle got call for which does not match the expected call format", zap.String(logFieldRequestMethod, method))
 		return
 	}
 
@@ -142,7 +142,7 @@ func (s *TChannelRouter) Handle(ctx context.Context, call *tchannel.InboundCall)
 	e, ok := s.endpoints[method]
 	s.RUnlock()
 	if !ok {
-		s.logger.Error("Handle got call for method which is not registered",
+		s.contextLogger.Error(ctx, "Handle got call for method which is not registered",
 			zap.String(logFieldRequestMethod, method),
 		)
 		return
@@ -167,10 +167,10 @@ func (s *TChannelRouter) Handle(ctx context.Context, call *tchannel.InboundCall)
 
 	var err error
 	c := &tchannelInboundCall{
-		call:     call,
-		endpoint: e,
-		logger:   s.logger,
-		scope:    s.scope.Tagged(scopeTags),
+		call:          call,
+		endpoint:      e,
+		contextLogger: s.contextLogger,
+		scope:         s.scope.Tagged(scopeTags),
 	}
 
 	c.start()